+++ Source/Processors/Visualization/SpikeDisplayCanvas.h
@@ -405,7 +405,7 @@ public:
 
 private:
 
-    void updateProjectionImage(uint16_t, uint16_t, uint16_t);
+    void updateProjectionImage(uint16_t, uint16_t, float);
 
     void calcWaveformPeakIdx(const SpikeObject&, int, int, int*, int*);
